project/requirements.txt
django==1.11.21
djangorestframework==3.7.7
django-extensions==2.1.6
gunicorn==19.9.0
cbor2==4.1.2
psycopg2-binary==2.8.3
whitenoise==4.1.2
requests==2.22.0
django-cors-headers==3.0.2